Force `rand_str` to return a non-numeric string
The way the string is created leaves a probability to have numeric strings
like "19283746" (even more probable when the string lentgh is short).

This can cause troubles when the string is used as an array key because
PHP cast it into an integer. And this caused some unit test to randomly
fail.